Cleaning up TestApplication
