+++ samples/TestApplication/Item.cs
@@ -1,0 +1,8 @@
+﻿namespace TestApplication
+{
+    internal class Item
+    {
+        public string Name { get; set; }
+        public string Value { get; set; }
+    }
+}
