+++ samples/TestApplication/Node.cs
@@ -1,0 +1,16 @@
+﻿using Perspex.Collections;
+
+namespace TestApplication
+{
+    internal class Node
+    {
+        public Node()
+        {
+            Children = new PerspexList<Node>();
+        }
+
+        public string Name { get; set; }
+        public PerspexList<Node> Children { get; set; }
+    }
+
+}
